+++ src/main/java/co/web/project/repository/IUsuariosRepository.java
@@ -3,6 +3,6 @@ package co.web.project.repository;
 import org.springframework.data.jpa.repository.*;
 import co.web.project.entity.Usuarios;
 
-public interface IUsuariosRepository extends JpaRepository<Usuarios, Long> {
+public interface IUsuariosRepository extends JpaRepository<Usuarios, Integer> {
 
 }
